moved pageNodeDropPlugin to separate file
